Adding Batch Processor test
