+++ Makefile
@@ -50,7 +50,7 @@ coverage-html:
 pre-commit:
 	pre-commit run --show-diff-on-failure
 
-pr: lint lint-docs mypy pre-commit test test-dependencies security-baseline complexity-baseline
+pr: lint lint-docs mypy pre-commit test security-baseline complexity-baseline
 
 build: pr
 	poetry build
